Move sdk to single entry point
