--- index.node.ts
@@ -1,2 +1,0 @@
-export * from './src/shared';
-export * from './src/node';
