+++ package.json
@@ -8,25 +8,6 @@
   "sideEffects": false,
   "license": "UNLICENSED",
   "exports": {
-    "./node": {
-      "import": {
-        "types": "./dist/node/index.d.ts",
-        "default": "./dist/node/index.js"
-      }
-    },
-    "./browser": {
-      "import": {
-        "types": "./dist/browser/index.d.ts",
-        "default": "./dist/browser/index.js"
-      }
-    },
-    ".": {
-      "browser": "./dist/browser/index.js",
-      "import": {
-        "types": "./dist/browser/index.d.ts",
-        "default": "./dist/browser/index.js"
-      }
-    },
     "./*": {
       "types": "./*.d.ts",
       "default": "./*.js"
