+++ src/Archaeologist.ts
