+++ src/Token.ts
