--- src/browser/index.ts
@@ -1,3 +1,0 @@
-import { WebSarcoClient } from './WebSarcoClient';
-export const sarco = new WebSarcoClient();
-export * from '../shared';
