+++ src/globalTypes.d.ts
