+++ src/index.ts
@@ -1,12 +1,53 @@
-export { WebSarcoClient } from './browser/WebSarcoClient';
-export * from './shared';
-export { sarco } from './browser/index';
-export type {
+import { WebSarcoClient } from './WebSarcoClient';
+
+export { NodeSarcoClient } from './NodeSarcoClient';
+export { WebSarcoClient };
+export const sarco = new WebSarcoClient();
+
+import { NEGOTIATION_SIGNATURE_STREAM } from './libp2p_node/p2pNodeConfig';
+import { goerliNetworkConfig, mainnetNetworkConfig, sepoliaNetworkConfig } from './networkConfig';
+import {
+  CallOptions,
+  RecoverPublicKeyErrorStatus,
+  RecoverPublicKeyResponse,
+  SarcoBundlrConfig,
+  SarcoNetworkConfig,
+} from './types';
+import {
+  ArchaeologistCurseNegotiationParams,
   ArchaeologistData,
   ArchaeologistEncryptedShard,
-  SarcophagusArchaeologist,
   ArchaeologistException,
-} from './shared/types/archaeologist';
-export type { SarcophagusData, SarcophagusDetails } from './shared/types/sarcophagi';
-export { SarcophagusState } from './shared/types/sarcophagi';
-export type { SarcoNetworkConfig } from './shared/types/index';
+  ArchaeologistExceptionCode,
+  ArchaeologistNegotiationResponse,
+  ArchaeologistNegotiationResult,
+  ArchaeologistProfile,
+  SarcophagusArchaeologist,
+  SarcophagusValidationError,
+} from './types/archaeologist';
+
+export type { CallOptions, RecoverPublicKeyResponse, SarcoBundlrConfig, SarcoNetworkConfig };
+
+export { RecoverPublicKeyErrorStatus };
+
+export {
+  ArchaeologistExceptionCode,
+  NEGOTIATION_SIGNATURE_STREAM,
+  SarcophagusValidationError,
+  goerliNetworkConfig,
+  mainnetNetworkConfig,
+  sepoliaNetworkConfig,
+};
+export type {
+  ArchaeologistCurseNegotiationParams,
+  ArchaeologistData,
+  ArchaeologistEncryptedShard,
+  ArchaeologistException,
+  ArchaeologistNegotiationResponse,
+  ArchaeologistNegotiationResult,
+  ArchaeologistProfile,
+  SarcophagusArchaeologist,
+};
+
+export { SarcophagusState } from './types/sarcophagi';
+export type { SarcophagusData, SarcophagusDetails } from './types/sarcophagi';
