--- src/node/index.ts
@@ -1,5 +1,0 @@
-import { NodeSarcoClientConfig } from './NodeSarcoClient';
-
-export { NodeSarcoClient } from './NodeSarcoClient';
-export type { NodeSarcoClientConfig };
-export * from '../shared';
