+++ src/types/archaeologist.ts
