+++ test/api/test-data.ts
@@ -1,5 +1,5 @@
 import { BigNumber, ethers } from 'ethers';
-import { ArchaeologistSettings, SarcophagusSettings } from '../../src/shared/helpers/validation';
+import { ArchaeologistSettings, SarcophagusSettings } from '../../src/helpers/validation';
 
 export const sarcoId = 'test-id';
 export const arweaveTxId = 'test-tx-id';
